feat: alterando a arquitura dos carros, update cars
