+++ backend/src/routes/cars.ts
@@ -3,7 +3,7 @@ import { Router } from "express";
 import allCarsController from "../controllers/cars/all-cars-controller";
 import createCarController from "../controllers/cars/create-car-controller";
 import deleteCarController from "../controllers/cars/deleteCar";
-import updateCarController from "../controllers/cars/updateCars";
+import updateCarController from "../controllers/cars/update-car-controller";
 
 const routeCar = Router();
